# server/swagger.yaml
openapi: 3.0.3
info:
  title: Rocko Backend API Docs
  version: 1.0.1
servers:
  - url: https://backend.dev.rocko.cloud
  - url: http://localhost:5000
  - url: https://backend.prod.rocko.cloud
...
paths:
...
  /alerts:
    get:
        tags:
          - alert
        summary: ""
        description: ""
        parameters:
          - name: loanId
            in: query
            description: ""
            required: false
            explode: true
            schema:
              type : integer
              format: int64
              minimum: 1
        responses:
          '200':
            description: successful operation
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/Alert'          
              application/xml:
                schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/Alert'
          '400':
            description: Invalid value
        security:
          - bearerAuth: [] 
  /addAlert:
    post:
        tags:
          - alert
        summary: ""
        description: ""
        requestBody:
         content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddAlert'
            application/xml:
              schema:
                $ref: '#/components/schemas/AddAlert'
            application/x-www-form-urlencoded:
              schema:
                $ref: '#/components/schemas/AddAlert'
        responses:
          '200':
            description: Alert successfully saved
          '400':
            description: Invalid status value
          '401':
            description: Unauthorized
        security:
          - bearerAuth: [] 
  /updateAlert: 
     post:
        tags:
          - alert
        summary: ""
        description: ""
        requestBody:
         content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateAlert'
            application/xml:
              schema:
                $ref: '#/components/schemas/UpdateAlert'
            application/x-www-form-urlencoded:
              schema:
                $ref: '#/components/schemas/UpdateAlert'
        responses:
          '200':
            description: Alert successfully updated
          '400':
            description: Invalid value
          '401':
            description: Unauthorized
        security:
          - bearerAuth: [] 
  /deleteAlert:
    post:
        tags:
          - alert
        summary: ""
        description: ""
        requestBody:
         content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAlert'
            application/xml:
              schema:
                $ref: '#/components/schemas/DeleteAlert'
            application/x-www-form-urlencoded:
              schema:
                $ref: '#/components/schemas/DeleteAlert'
        responses:
          '200':
            description: Alert successfully removed
          '400':
            description: Invalid value
          '401':
            description: Unauthorized
        security:
          - bearerAuth: [] 
  /average_apr:
    get:
        tags:
          - asset
        summary: ""
        description: ""
        parameters:
          - name: openDate
            in: query
            description: ""
            required: false
            explode: true
            schema:
              type : string
              example: month
          - name: network
            in: query
            description: ""
            required: false
            explode: true
            schema:
              type : string
              example: sepolia
        responses:
          '200':
            description: ""
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    average_apr:
                      type: number
          '400':
            description: Invalid value
  /average_reward_rate:
    get:
        tags:
          - asset
        summary: ""
        description: ""
        parameters:
          - name: network
            in: query
            description: ""
            required: false
            explode: true
            schema:
              type : string
              example: sepolia
        responses:
          '200':
            description: ""
            content:
              application/json:
                schema:
                  type: array
                  items:
                    type: object
                    properties:
                      average_reward_rate:
                        type: number
          '400':
            description: Invalid status value
  /reward_rate:
    get:
        tags:
          - asset
        summary: ""
        description: ""
        parameters:
          - name: network
            in: query
            description: ""
            required: false
            explode: true
            schema:
              type : string
              example: sepolia
        responses:
          '200':
            description: ""
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    borrow_reward_rate:
                      type: number
          '400':
            description: Invalid status value
...
  /cb-callback:
    get:
      tags:
        - coinbase
      summary: ""
      description: ""
      parameters:
        - name: code
          in: query
          description: ""
          required: false
          explode: true
          schema:
            type : string
            format: int64
            minimum: 1
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid status value
      security:
        - bearerAuth: [] 
  /coinbase-balance:
    get:
      tags:
        - coinbase
      summary: ""
      description: ""
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid status value
      security:
        - bearerAuth: [] 
components:
  schemas:
...
    Alert:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        loan_id:
          type: integer
          format: int64
          example: 21
        alert_type:
          type: string
          example: Collateral
          enum:
            - Collateral
            - APR
        alert_metric:
          type: string
          example: Below
          enum:
            - Below
            - Above
        alert_threshold:
          type: integer
          example: 15
        alert_email:
          type: integer
          example: 0
        alert_phone: 
          type: integer
          example: 1
        alert_repeat_secs:
          type: integer
          example: 0
        alert_once:
          type: integer
          example: 0
        active:
          type: boolean
          example: 1
        triggered:
          type: integer
          example: 1
        triggered_time:
          type: string
        user_id:
          type: integer
          example: 19
    AddAlert:
      type: object
      properties:
        loan_id:
          type: integer
          format: int64
          example: 21
        alert_type:
          type: string
          example: Collateral
          enum:
            - Collateral
            - APR
        alert_metric:
          type: string
          example: Below
          enum:
            - Below
            - Above
        alert_threshold:
          type: integer
          example: 15
        alert_email:
          type: integer
          example: 0
        alert_phone: 
          type: integer
          example: 1
        alert_repeat_secs:
          type: integer
          example: 0
        alert_once:
          type: integer
          example: 0
        active:
          type: boolean
          example: 1
    UpdateAlert:
      type: object
      properties:
        id: 
          type: integer
          example: 114
        alert_metric:
          type: string
          example: Below
          enum:
            - Below
            - Above
        alert_threshold:
          type: integer
          example: 15
        alert_email:
          type: integer
          example: 0
        alert_phone: 
          type: integer
          example: 1
        alert_repeat_secs:
          type: integer
          example: 0
        alert_once:
          type: integer
          example: 0
        active:
          type: boolean
          example: 1
    DeleteAlert: 
        type: object
        properties:
          id: 
            type: integer
            example: 114
          active:
            type: boolean
            example: 0
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT 